requirements.txt
PyPDF2
requests
google-generativeai
streamlit
python-dotenv